dev: Move label classes into new file
