+++ mealie/services/parser_services/openai/__init__.py
@@ -1,3 +1,4 @@
-from .parser import OpenAILabelMatcher, OpenAIParser
+from .match_labels import OpenAILabelMatcher
+from .parser import OpenAIParser
 
 __all__ = ["OpenAIParser", "OpenAILabelMatcher"]
